requirements.txt
aniso8601==9.0.1
APScheduler==3.10.4
async-generator==1.10
attrs==22.1.0
beautifulsoup4
blinker==1.7.0
bcrypt==4.2.0
certifi==2022.6.15
#cffi==1.16.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.7
dnspython==2.2.1
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
flask-swagger-ui==4.11.1
flask_jwt_extended==4.6.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
#numpy
#penpyxl==3.1.0
#outcome==1.2.0
#pandas
pycparser==2.21
PyJWT==2.8.0
pymongo==4.2.0
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
reportlab==4.2.2
requests==2.28.1
selenium==4.4.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.5
tabula==1.0.5
trio==0.21.0
trio-websocket==0.9.2
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.11
Werkzeug==2.2.2
wsproto==1.1.0
xlrd==2.0.1
XlsxWriter==3.0.3
xlutils==2.0.0
xlwt==1.3.0
zipp==3.8.1